Changed talk from 'tinitus' to 'experiments'
I'm hoping Chris is cool with this, my plan was to talk about a tinitus project I've been developing, but I've just not gathered enough research to do the subject justice.

However, I've just completed a project that uses WebRTC beyond the typical Skype-like demos, combining video with HTML5 canvas and the peer connection API, and have learnt tons that I'd love to share at jsconf.

So, I'm hoping Chris will accept this pull request, I'll buy him a beer or two (or rather *fetch him one of the free beers he's paid for*), and maybe give him a backrub and a cuddle if he does.